Finish up PostIcon / Card components
