+++ site/content/2021/how-i-built-this-(blog).mdx
@@ -7,12 +7,12 @@ author: Chris Carrick
 published: 2021-10-26T18:15:45.000Z
 tags:
   [
-    'react',
     'gatsby',
     'styled-components',
+    'mdx',
     'sanity',
     'contentful',
-    'mdx',
+    'react',
     'javascript',
     'typescript',
   ]
