+++ site/content/2021/meact-redux.mdx
@@ -5,7 +5,7 @@ featuredImage: ../../src/assets/images/code__laptop.jpg
 featuredIcon: redux
 author: Chris Carrick
 published: 2021-10-28T11:48:12.000Z
-tags: ['react', 'redux', 'rxjs', 'javascript', 'typescript']
+tags: ['redux', 'rxjs', 'react', 'javascript', 'typescript']
 ---
 
 These is your very first content with Contentful, pulled in JSON format using the [Content Delivery API](https://www.contentful.com/developers/docs/references/content-delivery-api/ 'Content Delivery API'). Content and presentation are now decoupled, allowing you to focus your efforts in building the perfect app.
