+++ site/content/2021/myled-components.mdx
@@ -5,7 +5,7 @@ featuredImage: ../../src/assets/images/keyboard.jpg
 featuredIcon: styled-components
 author: Chris Carrick
 published: 2021-10-29T23:07:13.000Z
-tags: ['react', 'styled-components', 'javascript', 'typescript']
+tags: ['styled-components', 'react', 'javascript', 'typescript']
 ---
 
 ## What are webhooks?
