--- site/src/components/ComboCard/index.js
@@ -1,1 +1,0 @@
-export * from './Card';
