--- site/src/components/DevIcon/index.js
@@ -1,1 +1,0 @@
-export * from './DevIcon';
