--- site/src/components/InlineCard/index.js
@@ -1,1 +1,0 @@
-export * from './InlineCard';
